Worked on Documentation and Compiler Bugfixes
