+++ src/commands.ts
@@ -1,20 +1,5 @@
-export interface commandsI{
-	[key: string]: (input:string)=>loginDetails
-}
-export interface loginDetails{
-	protocol: string,
-	host: string,
-	username: string,
-	password: string,
-	[key: string]: string
-}
-export const commands:commandsI = {
-	get: (input: string):loginDetails=>{
-		return {
-			protocol:"https",
-			host:"github.com",
-			username:"mockUser",
-			password:"mockPwd"
-		};
-	}
-}
+import get from "./commands/get"
+export default {
+	get,
+
+} as const
