+++ src/flags/config.ts
@@ -1,3 +1,3 @@
-export default async ()=>{
-
+export default async ():Promise<void>=>{
+	return
 }
